BINIUS-243: move gadget hints from binius-frontend to binius-circuits
The hints module mixed two unrelated things: the generic hint
infrastructure (the Hint trait, the registry, the id derivation) and
six concrete gadget hints that compute bignum and curve arithmetic.

The concrete hints only ever touch the public Hint trait plus
CircuitBuilder/Wire/Word, and every one of their real callers lives in
binius-circuits. So they never needed to sit in the frontend.

Keep the infrastructure in the frontend and move each concrete hint
next to the gadget that uses it:

- bignum/: BigUintDivideHint, BigUintModPowHint, ModDivideHint,
  ModInverseHint (all big-integer / modular arithmetic).
- secp256k1/: Secp256k1EndosplitHint (curve-specific).
- concat.rs: ByteVecConcatHint, inlined and made private since it has a
  single in-module caller.

Frontend tests can't depend on binius-circuits, so the six concrete
hint tests move with their hints. The const-prop-through-a-hint test is
rewritten to use a tiny local hint, since it exercises frontend
infrastructure rather than any concrete gadget.
