+++ crates/circuits/src/concat.rs
@@ -1,11 +1,62 @@
 // Copyright 2025 Irreducible Inc.
 use binius_core::{consts::WORD_SIZE_BYTES, word::Word};
-use binius_frontend::{CircuitBuilder, Wire, hints::ByteVecConcatHint};
+use binius_frontend::{CircuitBuilder, Wire, hints::Hint};
 
 use crate::{
 	fixed_byte_vec::{ByteVec, extract_const_range},
 	slice::{assert_slice_eq, slice},
 };
+
+/// Hint computing the concatenation of a list of fixed-capacity byte vectors.
+///
+/// Each input is described by a wire-count dimension `d_i` and contributes `d_i` data wires
+/// (8 bytes each, little-endian) plus one `len_bytes` wire. The hint reads the actual byte
+/// prefix of each input (per its `len_bytes`) and packs the concatenated bytes back into
+/// `sum(d_i)` little-endian output words, zero-padding any trailing space.
+struct ByteVecConcatHint;
+
+impl ByteVecConcatHint {
+	const fn new() -> Self {
+		Self
+	}
+}
+
+impl Hint for ByteVecConcatHint {
+	const NAME: &'static str = "binius.byte_vec_concat";
+
+	fn shape(&self, dimensions: &[usize]) -> (usize, usize) {
+		let total_data: usize = dimensions.iter().sum();
+		(total_data + dimensions.len(), total_data)
+	}
+
+	fn execute(&self, dimensions: &[usize], inputs: &[Word], outputs: &mut [Word]) {
+		let total_data: usize = dimensions.iter().sum();
+		let (data_wires, len_wires) = inputs.split_at(total_data);
+
+		let mut bytes = Vec::with_capacity(total_data * 8);
+		let mut cursor = 0;
+		for (&d, &len_word) in dimensions.iter().zip(len_wires) {
+			let words = &data_wires[cursor..cursor + d];
+			cursor += d;
+			let len = (len_word.as_u64() as usize).min(d * 8);
+			bytes.extend(
+				words
+					.iter()
+					.flat_map(|w| w.as_u64().to_le_bytes())
+					.take(len),
+			);
+		}
+
+		for (out, chunk) in outputs.iter_mut().zip(bytes.chunks(8)) {
+			let mut buf = [0u8; 8];
+			buf[..chunk.len()].copy_from_slice(chunk);
+			*out = Word(u64::from_le_bytes(buf));
+		}
+		for out in &mut outputs[bytes.len().div_ceil(8)..] {
+			*out = Word::ZERO;
+		}
+	}
+}
 
 /// Computes the concatenation of a list of [`ByteVec`]s as a new [`ByteVec`].
 ///
@@ -14,9 +65,9 @@ use crate::{
 /// - runtime length equal to the sum of the inputs' `len_bytes` values, and
 /// - `len_range` equal to the sum of the inputs' ranges (`sum(start)..sum(end)`).
 ///
-/// The output data wires are populated by [`ByteVecConcatHint`]; soundness is enforced by
-/// constraining each input's region of the output to equal that input's data. How that constraint
-/// is emitted depends on what is known at circuit-build time:
+/// The output data wires are populated by a prover-side concatenation hint; soundness is enforced
+/// by constraining each input's region of the output to equal that input's data. How that
+/// constraint is emitted depends on what is known at circuit-build time:
 ///
 /// - When a term's byte offset into the output is a compile-time constant (i.e. every preceding
 ///   term has a constant length), its region is extracted with constant shifts and masks (no
