+++ crates/circuits/src/secp256k1/endosplit.rs
@@ -17,19 +17,11 @@
 //! This is a hint - a deterministic computation that happens only on the prover side.
 //! The result should be additionally constrained by using bignum circuits to check that
 //! `k1 + λ k2 = k (mod n)`.
-//!
-//! The method used here comes straight from libsecp256k1, follow the link for derivation:
-//! <https://github.com/bitcoin-core/secp256k1/blob/master/src/scalar_impl.h#L92-L141>
 
 use binius_core::Word;
+use binius_frontend::{CircuitBuilder, Wire, hints::Hint, util::num_biguint_from_u64_limbs};
 use hex_literal::hex;
 use num_bigint::BigUint;
-
-use super::Hint;
-use crate::{
-	compiler::{CircuitBuilder, Wire},
-	util::num_biguint_from_u64_limbs,
-};
 
 pub struct Secp256k1EndosplitHint {
 	minus_b1: BigUint,
@@ -179,4 +171,55 @@ fn div_pow2_round(value: BigUint, shift: u64) -> BigUint {
 		BigUint::from(1usize)
 	};
 	(value >> shift) + increment
+}
+
+#[cfg(test)]
+mod tests {
+	use proptest::prelude::*;
+
+	use super::*;
+
+	proptest! {
+		#[test]
+		fn prop_secp256k1_endosplit(k in any::<[u64; 4]>()) {
+			let modulus = num_bigint::BigUint::from_bytes_be(
+				&hex_literal::hex!("fffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd0364141")
+			);
+			let lambda =  num_bigint::BigUint::from_bytes_be(
+				&hex_literal::hex!("5363ad4cc05c30e0a5261c028812645a122e22ea20816678df02967c1b23bd72")
+			);
+			let k_bignum = num_biguint_from_u64_limbs(k.iter());
+			prop_assume!(k_bignum < modulus);
+			prop_assume!(k_bignum > num_bigint::BigUint::ZERO);
+
+			let builder = CircuitBuilder::new();
+			let k = k.map(|limb| builder.add_constant_64(limb));
+			let (k1_neg, k2_neg, k1_abs, k2_abs) =
+				Secp256k1EndosplitHint::call(&builder, &k);
+
+			let circuit = builder.build();
+			let mut w = circuit.new_witness_filler();
+			circuit.populate_wire_witness(&mut w).unwrap();
+
+			let k1_abs_bignum = num_biguint_from_u64_limbs(k1_abs.iter().map(|&l| &w[l].0));
+			let k2_abs_bignum = num_biguint_from_u64_limbs(k2_abs.iter().map(|&l| &w[l].0));
+
+			assert!(k1_abs_bignum.bits() <= 128);
+			assert!(k2_abs_bignum.bits() <= 128);
+
+			let k1 = if w[k1_neg] != Word::ZERO {
+				&modulus - k1_abs_bignum
+			} else {
+				k1_abs_bignum
+			};
+
+			let k2 = if w[k2_neg] != Word::ZERO {
+				&modulus - k2_abs_bignum
+			} else {
+				k2_abs_bignum
+			};
+
+			assert_eq!((k1 + lambda * k2) % modulus, k_bignum);
+		}
+	}
 }
